Fix exception in Qt5-only environments
The DEFAULT_QT_API value in taurus custom settings was reset to pyqt4
in a recent commit. This introduces a bug in PyQt5-only environments
(unless QT_API is defined explicitly) because the search order of
bindings in taurus.external.qt assumes PyQt5 to be the default.
Adapt the search order accordingly.
Also related: change the fallback in case DEFAULT_QT_API is not set to
pyqt too.