Use weakref when enquing TangoAttribute.fireEvent jobs
Enquing hard reference to the TangoAttribute method as a job of the
thread pool may cause a life harder to the garbage collector when
TaurusSerializationMode.Serial serialization mode is in use.
Tests introduced in #847 demonstrates that. In this case the
TangoAttribute object creation enques its fireEvent jobs on the thread
pool, what implies a hard reference to this object. Until the thread pool
does not handle all the jobs, the object persist in the factory, even if there
are no external referrers - see #847. Furthermore, real events from the
Tango may arrive and add more jobs (more references) complicating the things
even more...

Use weakref (_BoundMethodWeakrefWithCall) so whenever there are no external
referrers the object can be destroyed by the garbage collector, even if there
are its jobs on the queue - no one is interested in handling them.
It was tested that a call to a dead reference by a worker thread does not
cause any exceptions.