Avoid redundant display of units in TaurusForm (Fixes #642)
TaurusForm uses TaurusValueLineEdit as a default write widget.
If the attribute is a numeric scalar, this shows the magnitude and the
units, which is redundant since the units are already displayed in the
units widget. Replace the default line edit in TarusValue by a
customized version that displays only the magnitude.

This is a (quick and dirty) fix for #642