Add auto-deploy to PyPI with Travis
Make Travis deploy to PyPI when a version tag (major.minor.patch) is
pushed. It uses an ad-hoc PyPI account (taurus_bot) which was given
Maintainer permisions in PyPI. The taurus_bot password is encrypted
with the taurus-org/taurus Travis public key so that Travis can decrypt
it for deployment.