Improve discovery of submodules by ModuleExplorer
ModuleExplorer.exploreModule discovers subpackages but fails to explore
submodules impleented as single files. This causes many classes to be
undocumented unless they are explicitly imported in some parent module.

Fix this by discovering single-file submodules (but filtering out those
which *are* already imported since those are not considered part of the
API)