Remove utf8 encodeing in JSONCodec
Taurus JSONCodec apart from JSON serialization encodes and decodes using
"utf-8". This in python 2 means that the API is based on strings and
in python 3 it is based on bytes.

In pure Python however json.dumps and json.loads works on strings
for both python 2 and 3.

Adapt Taurus JSONCodec to have the same behaviour than pure Python
library.

Solve  #945