Fix docs build in tox (and travis) travis
The sphinx build build was failing in py37-qt5-docs because of a
missing QT_QPA_PLATFORM envvar in the environment variable (as a
part of tox efforts to isolate environments, only a few variables
are passed to the environment). Fix it by declaring
QT_QPA_PLATFORM=offscreen in a setenv tox directive.