Avoid circular reference with TaurusDevice
The TaurusPollingTimer keeps device strong references
for creating the  weakref dictionary of polled attributes.
The removeAttribute does not cleanup the structures and
many of TaurusPollingTimer methods are not thread-safe.

Fix removeAttribute implementation to delete unreferenced
elements.
Protect the methods using the lock or making copy of the
iterated containers.

Fix #999