Improve handling of limits in QWheelEdit
QWheelEdit's current implementation does not properly distinguish
between the value limits imposed by the setMinValue / setMaxValue
methods and those imposed by the amount of digits available in the wheel
widget. Fix this by using a single validator (which is now shared by the
wheel and the editor) that properly combines both types of limits.