Add CI/CD debian packaging configuration (.gitlab-ci-alba.yml)
The official debian packaging of tarus is currently done in
salsa.debian.org , but at ALBA we are experimenting with a more
integrated CI/CD approach that auto-creates (unofficial) packages
for stretch, buster and sid on each upstream change. This requires
to add a configuration file to upstream (.gitlab-ci-alba.yml).
This file does nothing unless you explicitly configure your gitlab
instance to run on it.

TODO: generalise this file so that it is not dependent on ALBA's
infrastructure so that anyone can make use of it.