Add dummy JDBar support in TaurusJDrawGraphicsFactory
The Taurus JDraw viewer gets an exception if JDBar item is used in the
jdw file because TaurusJDrawGraphicsFactory does not support it.
Work around this by implementing a dummy support for JDBar.
Note that this is not proper support, but just a workaround
that displays a filled JDRectangle instead of a bar to allow parsing.