Revert unwanted change that made custom panels not modifiable
Panels created with `createCustomPanel` should be modifiable by the user
and contain the model in the config. This was removed in a previous
commit of this feature branch, but it is not wanted. Revert this
precise change.

Note: eventually a better solution would be to allow this to be chosen
by the user in the advanced options dialog, but until then we leave it
as originally.