Implicit optparse legacy support:
TaurusApplication implicitly uses a `optparse.OptionParser` instance
if it does not receive an explicit `cmd_line_parser` keyword argument.
This is inconvenient because it forces the user to explicitly pass
`cmd_line_parser=None` when using other mechanisms such as `click` or
`argparse` to parse CLI options.
Change this so that, by default cmd_line_parser=None is assumed if not
explicitly passed, but allow the user to revert to the old behaviour
by setting IMPLICIT_OPTPARSE=True in tauruscustomsettings.