Remove unneeded docker-compose configs
The SKA make file will try to pull all images even
if they are not used. Remove the .yml docker-compose
files that are not needed for CI.
Note: I leave the jive one because it may be useful
for for non-CI usage and it shares image with
tangotest (thus not increasing the download
size).