Allow failures in py2 tests
Quarantine the py2 tests. We are seeing them
finishing the test suite ok but showing problems
when *exiting* the pytest process (either
hanging or crashing) when run on Travis.

When run locally the crashes are a lot less
frequent.

This needs investigation (but we quarantine for
now to get some value from the rest of the CI).

Note: given that py2 support may be dropped soon,
this quarantine may be effective until removal.