LineEdit: enforce explicit disable
TaurusValueLineEdit auto enables/disables itself
based based on events. This precludes the user
from explicitly disable it. Fix by avoiding reenabling
if setEnabled(False) is called.
Note that auto-disabling on errors is still allowed
even if setEnabled(True) is called, since the widget
is not usable in case of error.