Fix #1113 (TangoAttribute not reenabling polling)
TangoAttribute calls `.disablePolling()` when
unsubscribing from Change events (this happens automatically
e.g. when removing its last listener). This is problematic
because it permanently disables the polling (e.g. it
won't e re-activated if a listener is lately added to the
same attribute). Prevent this by "deactivating" instead of
"disabling" the polling.